yolo2coco_segmentation/requirements.txt
Pillow===10.1.0
json5==0.9.14
argparse
